removeIdenticalField: BUG: Field separator determination incorrectly complains about -F ' ' being a regular expression when the first input line is separated by multiple spaces
With multiple spaces, the first branch doesn't apply, and the second one is only concerned about tab separation.
Change the condition so that the second branch is taken if (FS == " "); that's definitely not a regular expression, but the default. Then, just decide on whether space or tab is used as OFS. This will still produce an unexpected result (space output separation) if the first input line is separated by multiple tabs.