Cosmetics: Tweak the month label positions
So that they fit the start of each month; depending on the first weekday of the year, this can vary by 7 days.
01-Jan-2024 is a Monday, so charting starts right at the top-left cell.
01-Jan-2023 is a Sunday, so charting starts only at the bottom-left cell.

This command can graph all the month ranges (plus the possible positions for the last day of the year (with 2012 being a leap year (one more day) that started on a Sunday).

$ echonl "2023-"{01..12}"-01 12:00" "2023-"{01..12}"-01 12:00" "2023-"{01..12}"-01 12:00" "2023-"{01..12}"-01 12:00" \
	"2024-"{01..12}"-01 12:00" \
	"2024-12-31 12:00" "2024-12-31 12:00" "2012-12-31 12:00" \
	"2012-12-31 12:00" "2012-12-31 12:00" | \
    DISTRIBUTIONOVERTHEYEAR_EMPTY_CELL='^[[38;5;253m🭽^[[0m' distribution-over-the-year \
	--graph-field 1 --as green-center-boxes