BUG: Keeps iterating after spark error, exit status 99 only in d...-year
When data is split into multiple graphs, a spark error (likely due to a misnamed command-line argument) will be printed repeatedly.
Also, the command should exit with 99 if there's no valid data, but only distribution-over-the-year has the logic and it doesn't work there because the iteration will always succeed.
Consistently start with a status of 99 and change that to 0 on the first successful spark invocation. Exit immediately in case of a spark error.
If the tally does not gather any (valid) data (when the count array is empty), do not print the data, so that printDistribution() will break out of the loop and exit with 99.