ENH: Drop extracting the legends if start and end dates lie within one range, unless new --force-legends
The automatic adding of the legends is useful if multiple graphs are printed, but for a single graph it's mostly superfluous information. This also prevents the use of distribution-over-autorange as a drop-in replacement for the distribution-over-the-* commands.
Other clients might like the legends, or invoke the command repeatedly with different input ranges. Allow forcing the legend with --force-legends for these use cases.