distribution-over-autorange: ENH: Allow explicit range selection via --over
This allows the users of clients to choose a certain range, regardless of the input data; many clients use the autorange functionality to show relevant statistics depending on the input data range. By adding an option to force a certain range, clients do not have to implement that themselves, and they benefit from the enhancements and field selection logic of this wrapper.