distribution-over-the-year: FIX: --split-graph-field misleadingly uses different sparkline scales for each split graph
When the value ranges are more or less the same over the entire dataset, this is barely recognizable.
But in a "git logdistribution --authors" in the todo.txt-cli project, there was an initial flurry of authors in the first year, followed by a calm period of only one or two authors each day. However, the graph showed all shades of green in the first year and then black dots thereafter, which can't be right (it should be light green indicating the few authors).
The easy thing is demanding that the client supplies the --max parameter, but he'd have to extract that information on their own. (In the case of Git we could cheat a bit and pass the total count of all-time authors instead of going through the data, but that could potentially drastically distort the graphs.)
To obtain the maximum ourselves, we need two passes over the data, first to determine the value range, then to graph with those ranges.
I've extracted a generic withPreprocessedInput tool that encapsulates the implementation details of saving and restoring the streamed data.
In order to reuse most of the processing logic, the AWK script is split, with the generic part in renderWith(), and both existing tally() and new determineCountRange() supplying init() and printSection() template methods.

CHG: Don't assume min=0, but auto-range the minimum as well. Clients that want to compare to a baseline of 0 now need to pass --min 0 for that.