distribution-over-the-month: FIX: --split-graph-field misleadingly uses different sparkline scales for each split graph
Same approach as with distribution-over-the-year, but this variant surfaced additional aspects:
Here, the slot is just the day-of-month, so the values from different months are lumped together, leading to wrongly larger max values. (For *-over-year, because of the larger distribution over 365 values, this effect was less pronounced.)
determineCountRange() also needs to "delete counts" on each printSection(), like the original function.
Move the final output of the min/max command-line parameters out of it and into a second END section (there can be multiple). As that section needs to run after the original one in the template (that invokes printSection() a final time), renderWith() now takes two code blocks, pre- and appending to the template code.
--split-graph-field now also needs to be passed to the preprocessing stage; add it to inputDataClassificationArgs. That weakens the recursion guard; we need to check $action (which we now default to avoid duplicating the default action) now as well.