statistics: ENH: Add --unit-once variant of --unit
For short units (like "kB"), it's okay to repeat, but if it's a full word (like " characters"), this would be unwieldy.
Offer an option to just include the unit once, appended to the last value that's denoted in it (the last two are unitless counts).

System-info: Armbian 24.11.1 noble, aarch64
Platform-info: bash 5.2.21(1)-release, Bats 1.11.0